GitHub issue #38: update dirty textures WebGL
Issue: TextureAtlases not updating upon dirty being toggled

Fixed by adding a “dirty check” to the rendering pipeline directly
before draw calls.

Added TextureAtlas.refreshTextureGL method for more transparent texture
updates. This should be compatible with future additions to the engine,
for example, a form of atlas that stores both colour and normal data.

Standardised non-batch render calls to use identical process. The only
non-batch renderer at this time is in the WebGLParticlePlugin, which
does a lot of its own heavy lifting. We’re trying to standardise all
that stuff for future engine enhancements, so it will be easier to
write new renderers.