Implement Animation fix for negative time
Still re issue #184

In the race circumstance that the animation initialised on a negative frame count equal to the positive length, the code would set `frameCount=length`. This is an out-by-one error.

The erroneous code which set negative frame indices to nominally in-range positive indices read:

    this.length + this._frameIndex % this.length

The fix was to re-remainder the results:

    ( this.length + this._frameIndex % this.length ) % this.length