Fix `Timer` causing program to hang when `delay=0`.
Re issue # 187

`delay` now must be greater than 0, or it will remain at default (0.016 clock units) and report a helpful error message.

Also update `Clock.setTimeout` and `setInterval` to use `Clock.units` instead of `1000`, which is the default value of `units` but really shouldn't be taken for granted. Re issue #188