Always update graphics on `requestAnimationFrame()`.
Re issue #204

When running files through CocoonJS on Android platforms, occasional flickering might occur.

The flickering was brief, single-frame only. Its contents seemed to be mostly random bits out of video memory: repeated segments of the last frame, etc. Significantly, it appeared to correlate with maximum frame rate: when the fps hit max, a flicker happened.

Some investigation revealed that CocoonJS uses multiple frame buffers. If one was not drawn, it would be displayed anyway, potentially with bizarre results.

KiwiJS uses `requestAnimationFrame()` to draw each frame. Critically, it will SKIP a frame if the minimum frame duration has not been met. Because of time rounding or associated fluctuations, this may sometimes occur when the game is running very close to maximum frame rate.

When a frame is skipped, it seems to be drawn anyway. If its contents have been scrambled behind the scenes, this will result in flicker.

To fix this, we need to draw something to that frame's buffer.

I have implemented a change to the game loop. It now executes the rendering every frame; only non-graphics elements are skipped if the frame rate is high. This is unlikely to result in increased power draw on devices; as they are already hitting the max frame rate by definition, they will not have much extra work to do. Devices that do not hit max frame rate should not be affected.