Fix shader conflict bug
Re issue #210

Creating a shader after `State.create` has run and the render loop begins caused display to crash.

This was caused by newly-created shaders being pushed to the GPU, but not notifying the render loop, causing it to skip resetting the shader to the one it expects.

I've fixed this by setting `GLRenderManager._currentRenderer = null` at the start of every `render` pass, forcing it to reapply everything.