Change `Kiwi.Geom.Ray` to be an extension of `Kiwi.Geom.Line`. Rebuilt docs.
Re Github issue #170

This gives `Ray` access to `Line` methods, and allows it to more efficiently compute `Circle` intersections. These previously required the creation of a new Line or invocation of Line prototype methods.

Also corrected 'Kiwi.Geom.Intersect.rayToCircle()` to correctly assess ray directionality, allowing it to return sooner if the ray is not facing the circle.