Make WebGL render sequence more accessible. Deprecate private function `GLRenderManager._switchTexture()`.
Re issue #246

`GLRenderManager.renderBatch` was deriving most of its values from `batch[ 0 ].entity`. This meant it would ignore values set - or altered - in the `_sequence` array. I've upgraded the method to use the sequence data instead.

I've also deprecated `GLRenderManager._switchTexture()`. It took an entity as a parameter, which made it pretty difficult to implement the above changes.