Fix `Group.replaceChild()` behavior.
Re Github issue #262

The method ran some unnecessary processes, such as a removal from parent that was later duplicated in a method call.

In addition, the method set the parent of the replaced child to `null` for no reason, which could mess up later scene operations.

Method is now lighter and more reliable.