Fix `Group.clear()` to properly remove parents.
Re Github issue #263

`clear()` now calls `removeChildren()` with default parameters. That function more efficiently removes and de-parents members.

Also made some optimisations to `removeChild()` and `removeChildAt()`. This was supposed to support `clear()`, but turned out to be unnecessary. Formerly, `removeChildAt()` would use its index to derive the child, pass that to `removeChild()`, and then search for the index to do removal. This involves an entirely unnecessary search for an index that we already know. I've reversed the order - now, `removeChild()` will search for the index, but then it calls `removeChildAt()` to do the actual splicing.