Switch to compatibility with Cocoon.io.
Re Github issue #268

This change may not be backwards-compatible. However it is necessary to maintain compatibility with available Cocoon products.

Cocoon.io is now the only Cocoon platform available. It appears to implement the Canvas+ draw mode in a slightly different fashion (creating a normal canvas and setting `screencanvas=true` rather than creating a `screencanvas`). Accordingly, I've updated the library to be compatible with this new way of doing things.

This should not break anything, as there are no backwards-compatible Cocoon products left available.

Testing has been performed on several Android and iOS devices in a range of screen resolutions. Testbed app, built on Cocoon.io and fully readied for deployment, ran exactly as expected on all devices.